multiple bugs & UX suggestion fixes
