+++ src/css/base.scss
@@ -2811,6 +2811,10 @@ textarea,
     height: 60px !important;
 }
 
+.h-64 {
+    height: 64px !important;
+}
+
 .h-66 {
     height: 66px !important;
 }
